requirements.txt
openai==1.3.2
streamlit==1.28.2
pandas==2.1.3
python-dotenv==1.0.0